refactor(containers): split local rootfs outline view
